Vtk continue

+++ OpenMEEG/libs/OpenMEEG/include/interface.h
@@ -88,8 +88,8 @@ namespace OpenMEEG {
               bool          check(bool checked = false); ///< Check the global orientation
 
         /// \return the total number of the interface vertices
-        unsigned nb_vertices() const {
-            unsigned nb = 0;
+        size_t nb_vertices() const {
+            size_t nb = 0;
             for ( const_iterator omit = begin(); omit != end(); ++omit) {
                 nb += omit->mesh().nb_vertices();
             }
@@ -97,8 +97,8 @@ namespace OpenMEEG {
         }
         
         /// \return the total number of the interface triangles
-        unsigned nb_triangles() const {
-            unsigned nb = 0;
+        size_t nb_triangles() const {
+            size_t nb = 0;
             for ( const_iterator omit = begin(); omit != end(); ++omit) {
                 nb += omit->mesh().nb_triangles();
             }
